fix: symbol not found when building in baremental environment
