+++ runtime/internal/clite/c.go
@@ -21,7 +21,7 @@ import (
 )
 
 const (
-	LLGoPackage = "decl"
+	LLGoPackage = "link"
 )
 
 type (
