+++ runtime/internal/clite/debug/debug.go
@@ -1,4 +1,4 @@
-//go:build !wasm
+//go:build !wasm && !baremental
 
 package debug
