+++ runtime/internal/clite/debug/libunwind_linux.go
@@ -1,4 +1,4 @@
-//go:build linux
+//go:build linux && !baremental
 
 package debug
