+++ runtime/internal/lib/internal/syscall/unix/fcntl_baremental.go
@@ -1,0 +1,15 @@
+//go:build baremental
+
+package unix
+
+import (
+	_ "unsafe"
+)
+
+func Fcntl(fd int, cmd int, arg int) (int, error) {
+	return 0, nil
+}
+
+func fcntl(fd int, cmd int, arg int) (int, error) {
+	return 0, nil
+}
