+++ runtime/internal/lib/internal/syscall/unix/fcntl_unix.go
@@ -1,3 +1,5 @@
+//go:build !baremental
+
 package unix
 
 import (
