+++ runtime/internal/lib/runtime/maxprocs_baremental.go
@@ -1,0 +1,5 @@
+//go:build baremental
+
+package runtime
+
+func c_maxprocs() int32 { return 1 }
