+++ runtime/internal/lib/runtime/maxprocs_default.go
@@ -1,0 +1,11 @@
+//go:build !baremental
+
+package runtime
+
+const (
+	LLGoPackage = "link"
+	LLGoFiles   = "_wrap/runtime.c"
+)
+
+//go:linkname c_maxprocs C.llgo_maxprocs
+func c_maxprocs() int32
