+++ runtime/internal/runtime/rethrow_baremental.go
@@ -1,0 +1,6 @@
+//go:build baremental
+
+package runtime
+
+// Rethrow rethrows a panic.
+func Rethrow(link *Defer) {}
