+++ runtime/internal/runtime/z_rt.go
@@ -20,7 +20,6 @@ import (
 	"unsafe"
 
 	c "github.com/goplus/llgo/runtime/internal/clite"
-	"github.com/goplus/llgo/runtime/internal/clite/debug"
 	"github.com/goplus/llgo/runtime/internal/clite/pthread"
 	"github.com/goplus/llgo/runtime/internal/clite/setjmp"
 )
@@ -55,26 +54,6 @@ func Panic(v any) {
 	excepKey.Set(ptr)
 
 	Rethrow((*Defer)(c.GoDeferData()))
-}
-
-// Rethrow rethrows a panic.
-func Rethrow(link *Defer) {
-	if ptr := excepKey.Get(); ptr != nil {
-		if link == nil {
-			TracePanic(*(*any)(ptr))
-			debug.PrintStack(2)
-			c.Free(ptr)
-			c.Exit(2)
-		} else {
-			c.Siglongjmp(link.Addr, 1)
-		}
-	} else if link == nil && goexitKey.Get() != nil {
-		if pthread.Equal(mainThread, pthread.Self()) != 0 {
-			fatal("no goroutines (main called runtime.Goexit) - deadlock!")
-			c.Exit(2)
-		}
-		pthread.Exit(nil)
-	}
 }
 
 var (
